+++ refactoring_app_name/src/jp/co/sss/crud/exception/IllegalInputException.java
@@ -2,7 +2,7 @@ package jp.co.sss.crud.exception;
 
 public class IllegalInputException extends Exception{
 
-	IllegalInputException(String msg){
+	public IllegalInputException(String msg){
 		super(msg);
 	}
 }
