+++ refactoring_app_name/src/jp/co/sss/crud/io/EmployeeDeptIdReader.java
@@ -1,0 +1,7 @@
+package jp.co.sss.crud.io;
+
+public class EmployeeDeptIdReader {
+	public boolean isValid(String inputString) {
+		return inputString.matches("^[1-3１-３]{1}$");
+	}
+}
