+++ refactoring_app_name/src/jp/co/sss/crud/io/EmployeeGenderReader.java
@@ -1,0 +1,7 @@
+package jp.co.sss.crud.io;
+
+public class EmployeeGenderReader {
+	public boolean isValid(String inputString) {
+		return inputString.matches("^|[0129０１２９]{1}$");
+	}
+}
