+++ refactoring_app_name/src/jp/co/sss/crud/io/IConsoleReader.java
@@ -1,0 +1,5 @@
+package jp.co.sss.crud.io;
+
+public interface IConsoleReader {
+
+}
