+++ refactoring_app_name/src/jp/co/sss/crud/io/MenuNoReader.java
@@ -1,0 +1,8 @@
+package jp.co.sss.crud.io;
+
+public class MenuNoReader {
+
+	public boolean isValid(String inputString) {
+		return inputString.matches("^[1-7１-７]{1}$");
+	}
+}
