+++ refactoring_app_name/src/jp/co/sss/crud/main/MainSystem.java
@@ -6,7 +6,12 @@ import java.io.InputStreamReader;
 import java.sql.SQLException;
 import java.text.ParseException;
 
+import jp.co.sss.crud.exception.IllegalInputException;
 import jp.co.sss.crud.exception.SystemErrorException;
+import jp.co.sss.crud.io.EmployeeDeptIdReader;
+import jp.co.sss.crud.io.EmployeeGenderReader;
+import jp.co.sss.crud.io.EmployeeNameReader;
+import jp.co.sss.crud.io.MenuNoReader;
 import jp.co.sss.crud.service.EmployeeAllFindService;
 import jp.co.sss.crud.service.EmployeeDeleteService;
 import jp.co.sss.crud.service.EmployeeFindByDeptIdService;
@@ -44,6 +49,12 @@ public class MainSystem {
 		EmployeeUpdateService euService = new EmployeeUpdateService();
 		EmployeeDeleteService edService = new EmployeeDeleteService();
 		
+		MenuNoReader mnReader = new MenuNoReader();
+		EmployeeDeptIdReader ediReader = new EmployeeDeptIdReader();
+		EmployeeNameReader enReader = new EmployeeNameReader();
+		EmployeeGenderReader egReader = new EmployeeGenderReader();
+		
+		
 		int menuNo = ConstantValue.zero;
 
 		do {
@@ -61,6 +72,12 @@ public class MainSystem {
 			// メニュー番号の入力
 			String menuNoStr = br.readLine();
 			menuNo = Integer.parseInt(menuNoStr);
+			
+			//メニュー番号の入力チェック
+			if(!mnReader.isValid(menuNoStr)) {
+				System.out.println(ConstantMsg.MENU_NO_ERROR);
+				continue;
+			}
 
 			// 機能の呼出
 			switch (menuNo) {
@@ -68,7 +85,6 @@ public class MainSystem {
 				
 				try {
 					// 全件表示機能の呼出
-//					DBController.allSearch();
 					eafService.allSearch();
 					break;
 
@@ -83,14 +99,18 @@ public class MainSystem {
 				System.out.print(ConstantMsg.syainName);
 
 				// 検索機能の呼出
-//				DBController.syainSearch();
 				try {
 					efenService.syainSearch();
 					break;
+				} catch (IllegalInputException e) {
+					System.out.println(e.getMessage());
+					e.printStackTrace();
+					continue;
 				}catch (SystemErrorException e) {
 					System.out.println(e.getMessage()); 
 					e.printStackTrace(); 
 					break; 
+				
 				}
 
 
@@ -100,10 +120,16 @@ public class MainSystem {
 				String busyoId = br.readLine();
 
 				// 検索機能の呼出
-//				DBController.busyoSearch(busyoId);
 				try {
+					if(!(ediReader.isValid(busyoId))) {
+						throw new IllegalInputException(ConstantMsg.SYAIN_DEPTID_RANGE_ERROR);
+					}
 					efbdIService.busyoSearch(busyoId);
 					break;
+				} catch (IllegalInputException e) {
+					System.out.println(e.getMessage());
+					e.printStackTrace();
+					continue;
 				}catch (SystemErrorException e) {
 					System.out.println(e.getMessage()); 
 					e.printStackTrace(); 
@@ -123,10 +149,16 @@ public class MainSystem {
 				String busyoId2 = br.readLine();
 
 				try {
+					if(enReader.isValid()) {
+						
+					}
 					// 登録機能の呼出
-//					DBController.insert(emp_name, Seibetsu, birthday, busyoId2);
 					erService.insert(emp_name, Seibetsu, birthday, busyoId2);
 					break;
+				} catch (IllegalInputException e) {
+					System.out.println(e.getMessage());
+					e.printStackTrace();
+					continue;
 				}catch (SystemErrorException e) {
 					System.out.println(e.getMessage()); 
 					e.printStackTrace(); 
@@ -145,7 +177,6 @@ public class MainSystem {
 				Integer.parseInt(syainId);
 
 				// 更新機能の呼出
-//				DBController.update(syainId);
 				euService.update(syainId);
 				System.out.println(ConstantMsg.syainDataUpdate);
 
@@ -164,7 +195,6 @@ public class MainSystem {
 					System.out.print(ConstantMsg.syainDeleteIrai);
 	
 					// 削除機能の呼出
-//					DBController.delete();
 					edService.delete();
 					break;
 				}catch (SystemErrorException e) {
@@ -174,6 +204,7 @@ public class MainSystem {
 				}
 
 			}
+			
 		} while (menuNo != ConstantValue.seven);
 		System.out.println(ConstantMsg.systemEnd);
 	}
