+++ refactoring_app_name/src/jp/co/sss/crud/service/EmployeeDeleteService.java
@@ -1,6 +1,6 @@
 package jp.co.sss.crud.service;
 
-import jp.co.sss.crud.db.IEmployeeDAO;
+import jp.co.sss.crud.db.EmployeeDAO;
 import jp.co.sss.crud.exception.SystemErrorException;
 import jp.co.sss.crud.util.ConstantMsg;
 
@@ -9,7 +9,7 @@ public class EmployeeDeleteService {
 	public void delete() throws SystemErrorException{
 		try {
 			ConstantMsg c = new ConstantMsg();
-			IEmployeeDAO employeeDAO = new IEmployeeDAO();
+			EmployeeDAO employeeDAO = new EmployeeDAO();
 	
 			employeeDAO.delete();
 		} catch (Exception e) {
